test(repos): isolate runtime fallback proofs
