+++ src/renderer/src/store/slices/repos.test.ts
@@ -9,18 +9,6 @@ import {
 } from '../../runtime/runtime-compatibility-test-fixture'
 import { clearRuntimeCompatibilityCacheForTests } from '../../runtime/runtime-rpc-client'
 
-const toastError = vi.hoisted(() => vi.fn())
-const toastInfo = vi.hoisted(() => vi.fn())
-const toastSuccess = vi.hoisted(() => vi.fn())
-
-vi.mock('sonner', () => ({
-  toast: {
-    error: toastError,
-    info: toastInfo,
-    success: toastSuccess
-  }
-}))
-
 const localRepo: Repo = {
   id: 'local-repo',
   path: '/local',
@@ -83,9 +71,6 @@ beforeEach(() => {
   ptyKill.mockReset()
   runtimeEnvironmentCall.mockReset()
   runtimeEnvironmentTransportCall.mockReset()
-  toastError.mockReset()
-  toastInfo.mockReset()
-  toastSuccess.mockReset()
   runtimeEnvironmentTransportCall.mockImplementation((args: RuntimeEnvironmentCallRequest) => {
     return createCompatibleRuntimeStatusResponseIfNeeded(args) ?? runtimeEnvironmentCall(args)
   })
@@ -256,202 +241,6 @@ describe('repo slice runtime routing', () => {
     })
     expect(reposAdd).not.toHaveBeenCalled()
     expect(reposPickFolder).not.toHaveBeenCalled()
-  })
-
-  it('blocks wrong-host runtime fallback', async () => {
-    runtimeEnvironmentCall.mockImplementation((request: RuntimeEnvironmentCallRequest) => {
-      const { method } = request
-      const params = (request as { params?: unknown }).params
-      if (method === 'repo.add') {
-        return {
-          id: 'rpc-add-git',
-          ok: false,
-          error: { code: 'repo.invalid', message: 'Not a valid git repository' },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'projectGroup.create') {
-        return {
-          id: 'rpc-create-status-scope',
-          ok: true,
-          result: {
-            group: {
-              id: 'status-scope-1',
-              name: 'Path status check',
-              parentPath: (params as { parentPath: string }).parentPath,
-              parentGroupId: null,
-              createdFrom: 'manual',
-              tabOrder: 0,
-              isCollapsed: false,
-              color: null,
-              createdAt: 1,
-              updatedAt: 1
-            }
-          },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'folderWorkspace.getPathStatus') {
-        return {
-          id: 'rpc-path-status',
-          ok: true,
-          result: {
-            status: {
-              path: '/Users/me/GitHub/travel-hub',
-              exists: false,
-              reason: 'missing'
-            }
-          },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'projectGroup.delete') {
-        return {
-          id: 'rpc-delete-status-scope',
-          ok: true,
-          result: { deleted: true },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      throw new Error(`Unexpected runtime method ${method}`)
-    })
-    const store = createTestStore()
-    store.setState({
-      settings: { activeRuntimeEnvironmentId: 'env-1' } as never,
-      runtimeEnvironments: [{ id: 'env-1', name: 'Remote Mac' }] as never
-    })
-
-    await expect(
-      store.getState().addRepoPath('/Users/me/GitHub/travel-hub', 'git')
-    ).resolves.toBeNull()
-
-    expect(store.getState().activeModal).not.toBe('confirm-non-git-folder')
-    expect(runtimeEnvironmentCall).toHaveBeenCalledWith({
-      selector: 'env-1',
-      method: 'folderWorkspace.getPathStatus',
-      params: { scope: 'project-group', projectGroupId: 'status-scope-1' },
-      timeoutMs: 15_000
-    })
-    expect(runtimeEnvironmentCall).toHaveBeenCalledWith({
-      selector: 'env-1',
-      method: 'projectGroup.delete',
-      params: { groupId: 'status-scope-1' },
-      timeoutMs: 15_000
-    })
-    expect(toastError).toHaveBeenCalledWith(
-      'Cannot open folder on selected runtime',
-      expect.objectContaining({
-        description: expect.stringContaining('Remote Mac')
-      })
-    )
-  })
-
-  it('keeps runtime folder fallback on the checked host', async () => {
-    const folderRepo: Repo = {
-      ...remoteRepo,
-      id: 'runtime-folder',
-      path: '/srv/non-git',
-      displayName: 'non-git',
-      kind: 'folder'
-    }
-    runtimeEnvironmentCall.mockImplementation((request) => {
-      const { selector, method, params } = request as {
-        selector: string
-        method: string
-        params?: unknown
-      }
-      if (method === 'repo.add' && (params as { kind?: string }).kind === 'git') {
-        return {
-          id: 'rpc-add-git',
-          ok: false,
-          error: { code: 'repo.invalid', message: 'Not a valid git repository' },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'repo.add' && (params as { kind?: string }).kind === 'folder') {
-        return {
-          id: 'rpc-add-folder',
-          ok: true,
-          result: { repo: folderRepo },
-          _meta: { runtimeId: `runtime-${selector}` }
-        }
-      }
-      if (method === 'projectGroup.create') {
-        return {
-          id: 'rpc-create-status-scope',
-          ok: true,
-          result: {
-            group: {
-              id: 'status-scope-2',
-              name: 'Path status check',
-              parentPath: (params as { parentPath: string }).parentPath,
-              parentGroupId: null,
-              createdFrom: 'manual',
-              tabOrder: 0,
-              isCollapsed: false,
-              color: null,
-              createdAt: 1,
-              updatedAt: 1
-            }
-          },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'folderWorkspace.getPathStatus') {
-        return {
-          id: 'rpc-path-status',
-          ok: true,
-          result: {
-            status: {
-              path: '/srv/non-git',
-              exists: true
-            }
-          },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      if (method === 'projectGroup.delete') {
-        return {
-          id: 'rpc-delete-status-scope',
-          ok: true,
-          result: { deleted: true },
-          _meta: { runtimeId: 'runtime-remote' }
-        }
-      }
-      throw new Error(`Unexpected runtime method ${method}`)
-    })
-    const store = createTestStore()
-    store.setState({
-      settings: { activeRuntimeEnvironmentId: 'env-1' } as never,
-      fetchWorktrees: vi.fn().mockResolvedValue(undefined) as never
-    })
-
-    await expect(store.getState().addRepoPath('/srv/non-git', 'git')).resolves.toBeNull()
-
-    expect(store.getState().activeModal).toBe('confirm-non-git-folder')
-    expect(store.getState().modalData).toEqual({
-      folderPath: '/srv/non-git',
-      runtimeEnvironmentId: 'env-1'
-    })
-
-    store.setState({ settings: { activeRuntimeEnvironmentId: 'env-2' } as never })
-    await expect(
-      store.getState().addNonGitFolder('/srv/non-git', { runtimeEnvironmentId: 'env-1' })
-    ).resolves.toEqual({ ...folderRepo, executionHostId: 'runtime:env-1' })
-
-    expect(runtimeEnvironmentCall).toHaveBeenCalledWith({
-      selector: 'env-1',
-      method: 'repo.add',
-      params: { path: '/srv/non-git', kind: 'folder' },
-      timeoutMs: 15_000
-    })
-    expect(runtimeEnvironmentCall).not.toHaveBeenCalledWith(
-      expect.objectContaining({
-        selector: 'env-2',
-        method: 'repo.add',
-        params: { path: '/srv/non-git', kind: 'folder' }
-      })
-    )
   })
 
   it('sets up a project on a local host through the project setup API', async () => {
